feat(kubelet): 加入重启容错机制

+++ kubelet/cmd/server.go
@@ -39,9 +39,13 @@ type Empty struct{}
 
 var server kubeletServer = kubeletServer{}
 
+func generateServerName() string {
+	return "node-" + utils.GetLocalIP()
+}
+
 func serverInit() {
 	uid := uuid.NewString()
-	name := "node-" + uid[:6]
+	name := generateServerName()
 	server.Node = apiobjects.Node{
 		// TypeMeta: apiobjects.TypeMeta{
 		// 	ApiVersion: global.ApiVersion,
@@ -173,10 +177,17 @@ func healthReport() {
 		_, ok := server.Pods[ref.GetObjectPath()]
 		if !ok {
 			utils.Warn("kubelet:healthReport running pod not in kubelet internal list: ", ref.GetObjectPath())
-
-			utils.Info("kubelet:healthReport deleting pod: ", ref.GetObjectPath())
-			cri.DeletePod(podStatus.Status.Id)
-			continue
+			// Check if the pod is in the cluster
+			clusterPod, err := internal.GetPodByPath(ref.GetObjectPath())
+			if err == nil {
+				utils.Warn("kubelet:healthReport pod not in kubelet internal list but in cluster: ", ref.GetObjectPath())
+				// Add the pod to kubelet internal list
+				server.Pods[ref.GetObjectPath()] = clusterPod
+			} else {
+				// CHECK: 考虑不同的出错可能
+				utils.Warn("kubelet:healthReport pod not in kubelet internal list and not in cluster: ", ref.GetObjectPath())
+				cri.DeletePod(podStatus.Status.Id)
+			}
 		}
 	}
